+++ 1_DAL/IRepositories/IProductDetailRepository.cs
