+++ 1_DAL/IRepositories/ITypeTableOfContents.cs
