+++ 1_DAL/IRepositories/IreceiptDetailRepository.cs
