+++ 3_PL/3_PL.csproj
@@ -22,4 +22,8 @@
     <ProjectReference Include="..\2_BUS\2_BUS.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="View\" />
+  </ItemGroup>
+
 </Project>
